requirements.txt
pyzmq
pillow
opencv-contrib-python
numpy
pygame
readchar